// src/api/generated/model/createOrdContractTemplateRequestDtoDateEndStrategy.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Стратегия даты окончания. none = без даты; fixed = fixedDateEnd; offsetDays = date + offsetDays
 */
export type CreateOrdContractTemplateRequestDtoDateEndStrategy = typeof CreateOrdContractTemplateRequestDtoDateEndStrategy[keyof typeof CreateOrdContractTemplateRequestDtoDateEndStrategy];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateOrdContractTemplateRequestDtoDateEndStrategy = {
  none: 'none',
  fixed: 'fixed',
  offsetDays: 'offsetDays',
} as const;