// src/api/generated/model/createOrdContractTemplateRequestDtoDateStrategy.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Стратегия даты заключения. today = берётся текущая дата; fixed = используется fixedDate
 */
export type CreateOrdContractTemplateRequestDtoDateStrategy = typeof CreateOrdContractTemplateRequestDtoDateStrategy[keyof typeof CreateOrdContractTemplateRequestDtoDateStrategy];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateOrdContractTemplateRequestDtoDateStrategy = {
  today: 'today',
  fixed: 'fixed',
} as const;