// src/api/generated/model/createOrdContractTemplateRequestDtoFlagsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

export type CreateOrdContractTemplateRequestDtoFlagsItem = typeof CreateOrdContractTemplateRequestDtoFlagsItem[keyof typeof CreateOrdContractTemplateRequestDtoFlagsItem];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateOrdContractTemplateRequestDtoFlagsItem = {
  vat_included: 'vat_included',
  contractor_is_creatives_reporter: 'contractor_is_creatives_reporter',
  agent_acting_for_publisher: 'agent_acting_for_publisher',
  is_charge_paid_by_agent: 'is_charge_paid_by_agent',
} as const;